refactor: piece 관련 테스트 패키지 이동
