+++ src/test/java/model/piece/ElephantTest.java
@@ -1,4 +1,4 @@
-package model;
+package model.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
