+++ src/test/java/model/piece/GuardTest.java
@@ -1,6 +1,4 @@
-package model;
-
-import java.util.List;
+package model.piece;
 
 import model.piece.Guard;
 import model.piece.Piece;
@@ -36,9 +34,9 @@ public class GuardTest {
 
         // When & Then
         assertThat(guard.calculateRouteToDestination(destination))
-                .isEqualTo(List.of(new Position(5, 4)));
+                .containsExactlyInAnyOrder(new Position(5, 4));
         assertThat(guard.calculateRouteToDestination(destination2))
-                .isEqualTo(List.of(new Position(5, 6)));
+                .containsExactlyInAnyOrder(new Position(5, 6));
     }
 
     @Test
@@ -51,13 +49,13 @@ public class GuardTest {
 
         // When & Then
         assertThat(guard.calculateRouteToDestination(destination))
-                .isEqualTo(List.of(new Position(4, 4)));
+                .containsExactlyInAnyOrder(new Position(4, 4));
         assertThat(guard.calculateRouteToDestination(destination2))
-                .isEqualTo(List.of(new Position(4, 6)));
+                .containsExactlyInAnyOrder(new Position(4, 6));
         assertThat(guard.calculateRouteToDestination(destination3))
-                .isEqualTo(List.of(new Position(6, 4)));
+                .containsExactlyInAnyOrder(new Position(6, 4));
         assertThat(guard.calculateRouteToDestination(destination4))
-                .isEqualTo(List.of(new Position(6, 6)));
+                .containsExactlyInAnyOrder(new Position(6, 6));
     }
 
     @Test
@@ -68,9 +66,9 @@ public class GuardTest {
 
         // When & Then
         assertThat(guard.calculateRouteToDestination(destination))
-                .isEqualTo(List.of(new Position(4, 5)));
+                .containsExactlyInAnyOrder(new Position(4, 5));
         assertThat(guard.calculateRouteToDestination(destination2))
-                .isEqualTo(List.of(new Position(6, 5)));
+                .containsExactlyInAnyOrder(new Position(6, 5));
     }
 
     @Test
